requirements.txt
python-dotenv
lxml
bs4
unstructured
beautifulsoup4
pypdf
faiss-cpu
openai
langchain-openai
certifi
langchain
python-telegram-bot
requests
jq
